:truck: [FE] 폴더 변경
