--- frontend/src/babel.config.js
@@ -1,6 +1,0 @@
-module.exports = {
-    presets: [
-      '@babel/preset-env',   // 최신 JavaScript 기능을 위해
-      '@babel/preset-react'  // JSX 지원을 위해
-    ],
-  };
