--- frontend/src/components/Sidebar.css
@@ -1,67 +1,0 @@
-/* src/components/Sidebar.css */
-.dashboard {
-  font-family: "Roboto", sans-serif;
-  font-weight: 900;
-  font-style: italic;
-  font-size: 36px;
-
-  margin-top: 5%;
-}
-
-.sidebar-container {
-  display: flex;
-  flex-direction: column;
-}
-
-.sidebar ul {
-  list-style-type: none;
-  padding: 0;
-}
-
-.sidebar li {
-  font-weight: 600;
-  margin: 10px 0;
-}
-
-.sidebar a {
-  text-decoration: none;
-  color: #333;
-  transition: color 0.5s ease;
-
-}
-
-.admin-menu li a {
-  color: rgb(94, 176, 179);
-}
-
-@keyframes hover {
-  0% { color: #333; }
-  50% { color: #ebebeb; }
-  100% { color: #333; }
-}
-
-.sidebar a:hover {
-  animation: hover 2s infinite;
-}
-
-/* 채팅 css */
-
-.chat-wrapper {
-  display: flex;
-  justify-content: center;
-  box-sizing: border-box;
-  /* width: 230px; */
-  margin-top: auto; 
-}
-
-.chat-container {
-  width: 100%;
-  height: 400px;
-  border: 2px solid #c0c0c0;
-  border-radius: 5px;
-}
-
-.chat-header {
-  border-bottom: 2px solid #c0c0c0;
-  
-}
