+++ frontend/src/components/modal/Modal.css
@@ -13,9 +13,9 @@
   
   .modal-box {
     background: white;
-    padding: 20px;
-    border-radius: 8px;
-    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
+    padding: 1.25rem;
+    border-radius: 0.5rem;
+    box-shadow: 0 0.125rem 0.625rem rgba(0, 0, 0, 0.1);
   }
   
   .modal-content {
@@ -25,15 +25,15 @@
   }
   
   .modal-content p {
-    margin-bottom: 20px;
+    margin-bottom: 1.25rem;
   }
   
   .modal-content button {
-    padding: 10px 20px;
+    padding: 0.625rem 1.25rem;
     border: none;
     background-color: #007bff;
     color: white;
-    border-radius: 4px;
+    border-radius: 0.25rem;
     cursor: pointer;
   }
