+++ frontend/src/components/navbar/Navbar.css
@@ -1,6 +1,6 @@
 .navmenu {
     display: flex;
-    gap: 20px;
+    gap: 1.25rem;
 }
 
 .navmenu a {
